Loads a mesh
